requirements_310.txt
accelerate==1.12.0
aiohttp==3.13.2
attrs==25.4.0
certifi>=2023.5
cffi>=1.15.1
chardet>=5.2.0
charset-normalizer>=3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography>=36.0
datasets==4.4.1
dill==0.4.0
fastapi==0.121.3
filelock==3.14.0
frozenlist==1.8.0
fsspec>=2023.1
greenlet==3.2.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
Levenshtein==0.27.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
networkx>=3.0,<3.5
numpy>=1.23,<1.26
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
pandas>=2.0,<2.5
pillow>=9.5,<12
protobuf>=3.20,<6
psutil>=5.9,<8
pycparser>=2.23
pycryptodome>=3.16,<4
pydantic>=2.0,<2.13
Pygments==2.19.2
python-dateutil>=2.8,<2.9.3
python-multipart>=0.0.20
pytz>=2022,<2024
pywin32>=303,<312
PyYAML>=6.0,<7
regex<2025
requests>=2.28,<3.0
rich>=13.4,<14
safetensors>=0.7.0
setuptools>=60,<66
shapely>=2.0,<2.2
six>=1.17,<2
SQLAlchemy>=2.0,<2.1
sympy>=1.14,<1.15
tabulate>=0.9,<1
tokenizers>=0.22,<0.23
torch==2.9.1
torchvision==0.24.1
tqdm>=4.65,<5
transformers>=4.57,<4.60
typing_extensions>=4.15,<5
ujson>=5.11,<6
urllib3>=1.26,<2
uvicorn==0.38.0
wcwidth==0.2.14
xxhash>=3.6,<4
yarl>=1.22,<2
